Refactor NotesList and Note to a separated file
